reorganize(info): Split test suite into thematic files for better organization
Split the test suite into focused thematic files:
- edit_test.go: InfoFile editing operations (Add, Remove, Update)
- validation_test.go: InfoFileSet validation and cleaning operations
- multifile_test.go: InfoFileSet gather/distribute and utility operations
- api_fs_test.go: API filesystem integration tests
- Updated existing files to remove duplicated coverage and focus on core functionality

All 340 tests pass with maintained coverage. Improved test isolation and separation of concerns between pure unit tests and filesystem integration tests.
