+++ treex/info/info_test.go
@@ -3,10 +3,12 @@ package info
 import (
 	"testing"
 
-	"github.com/spf13/afero"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
+
+// TestInfoFile_Creation tests InfoFile creation and basic parsing
+// Core InfoFile functionality tests that don't duplicate edit operations
 
 func TestInfoFile_NewInfoFile(t *testing.T) {
 	content := `# This is a comment
@@ -90,72 +92,6 @@ file3.txt Annotation 3`
 	assert.True(t, paths["file3.txt"])
 }
 
-func TestInfoFile_UpdateAnnotationForPath(t *testing.T) {
-	content := `file1.txt Original annotation
-file2.txt Another annotation`
-
-	infoFile := NewInfoFile("/test/.info", content)
-
-	// Update existing annotation
-	result := infoFile.UpdateAnnotationForPath("file1.txt", "Updated annotation")
-	assert.True(t, result)
-
-	ann := infoFile.GetAnnotationForPath("file1.txt")
-	assert.Equal(t, "Updated annotation", ann.Annotation)
-
-	// Check that the line was updated
-	assert.Equal(t, "file1.txt Updated annotation", infoFile.Lines[0].Raw)
-
-	// Try to update non-existent annotation
-	result = infoFile.UpdateAnnotationForPath("nonexistent.txt", "New annotation")
-	assert.False(t, result)
-}
-
-func TestInfoFile_RemoveAnnotationForPath(t *testing.T) {
-	content := `file1.txt Annotation 1
-file2.txt Annotation 2
-file3.txt Annotation 3`
-
-	infoFile := NewInfoFile("/test/.info", content)
-
-	// Remove existing annotation
-	result := infoFile.RemoveAnnotationForPath("file2.txt")
-	assert.True(t, result)
-
-	assert.False(t, infoFile.HasAnnotationForPath("file2.txt"))
-	assert.Len(t, infoFile.annotations, 2)
-
-	// Check that the line was marked as removed
-	assert.Equal(t, LineTypeMalformed, infoFile.Lines[1].Type)
-	assert.Equal(t, "removed", infoFile.Lines[1].ParseError)
-
-	// Try to remove non-existent annotation
-	result = infoFile.RemoveAnnotationForPath("nonexistent.txt")
-	assert.False(t, result)
-}
-
-func TestInfoFile_AddAnnotationForPath(t *testing.T) {
-	content := `file1.txt Existing annotation`
-
-	infoFile := NewInfoFile("/test/.info", content)
-
-	// Add new annotation
-	result := infoFile.AddAnnotationForPath("file2.txt", "New annotation")
-	assert.True(t, result)
-
-	assert.True(t, infoFile.HasAnnotationForPath("file2.txt"))
-	ann := infoFile.GetAnnotationForPath("file2.txt")
-	assert.Equal(t, "New annotation", ann.Annotation)
-
-	// Check that a new line was added
-	assert.Len(t, infoFile.Lines, 2)
-	assert.Equal(t, "file2.txt New annotation", infoFile.Lines[1].Raw)
-
-	// Try to add duplicate annotation
-	result = infoFile.AddAnnotationForPath("file1.txt", "Duplicate annotation")
-	assert.False(t, result)
-}
-
 func TestInfoFile_IsEmpty(t *testing.T) {
 	// Empty file
 	emptyFile := NewInfoFile("/test/.info", "")
@@ -191,102 +127,4 @@ file1.txt Annotation 1
 # Another comment`
 
 	assert.Equal(t, expected, result)
-}
-
-func TestInfoFile_DuplicatePaths(t *testing.T) {
-	content := `file1.txt First annotation
-file1.txt Second annotation
-file2.txt Valid annotation`
-
-	infoFile := NewInfoFile("/test/.info", content)
-
-	// Should only have 2 annotations (first occurrence wins)
-	assert.Len(t, infoFile.annotations, 2)
-	assert.True(t, infoFile.HasAnnotationForPath("file1.txt"))
-	assert.True(t, infoFile.HasAnnotationForPath("file2.txt"))
-
-	// First annotation should win
-	ann := infoFile.GetAnnotationForPath("file1.txt")
-	assert.Equal(t, "First annotation", ann.Annotation)
-
-	// Second line should be marked as malformed
-	assert.Equal(t, LineTypeMalformed, infoFile.Lines[1].Type)
-	assert.Equal(t, "duplicate path (first occurrence wins)", infoFile.Lines[1].ParseError)
-}
-
-func TestInfoFileSetLoader_LoadSingleInfoFile(t *testing.T) {
-	fs := afero.NewMemMapFs()
-	infoFS := NewAferoInfoFileSystem(fs)
-	loader := NewInfoFileSetLoader(infoFS)
-
-	// Create a test .info file
-	content := `file1.txt Annotation 1
-file2.txt Annotation 2`
-
-	err := afero.WriteFile(fs, "/test/.info", []byte(content), 0644)
-	require.NoError(t, err)
-
-	// Load the file
-	infoFile, err := loader.LoadSingleInfoFile("/test/.info")
-	require.NoError(t, err)
-
-	assert.Equal(t, "/test/.info", infoFile.Path)
-	assert.Len(t, infoFile.annotations, 2)
-	assert.True(t, infoFile.HasAnnotationForPath("file1.txt"))
-	assert.True(t, infoFile.HasAnnotationForPath("file2.txt"))
-}
-
-func TestInfoFileSetLoader_LoadFromPath(t *testing.T) {
-	fs := afero.NewMemMapFs()
-	infoFS := NewAferoInfoFileSystem(fs)
-	loader := NewInfoFileSetLoader(infoFS)
-
-	// Create test .info files
-	err := afero.WriteFile(fs, "/project/.info", []byte("root.txt Root annotation"), 0644)
-	require.NoError(t, err)
-
-	err = fs.MkdirAll("/project/sub", 0755)
-	require.NoError(t, err)
-
-	err = afero.WriteFile(fs, "/project/sub/.info", []byte("sub.txt Sub annotation"), 0644)
-	require.NoError(t, err)
-
-	// Load all files
-	infoFileSet, err := loader.LoadFromPath("/project")
-	require.NoError(t, err)
-
-	infoFiles := infoFileSet.GetFiles()
-	assert.Len(t, infoFiles, 2)
-
-	// Check that both files were loaded
-	paths := make(map[string]bool)
-	for _, infoFile := range infoFiles {
-		paths[infoFile.Path] = true
-	}
-	assert.True(t, paths["/project/.info"])
-	assert.True(t, paths["/project/sub/.info"])
-}
-
-func TestInfoFileSetWriter_WriteSingleInfoFile(t *testing.T) {
-	fs := afero.NewMemMapFs()
-	infoFS := NewAferoInfoFileSystem(fs)
-	writer := NewInfoFileSetWriter(infoFS)
-
-	// Create directory first
-	err := fs.MkdirAll("/test", 0755)
-	require.NoError(t, err)
-
-	// Create an InfoFile with some content to avoid empty string issue
-	infoFile := NewInfoFile("/test/.info", "# Comment")
-	infoFile.AddAnnotationForPath("file1.txt", "Test annotation")
-
-	// Write to disk
-	err = writer.WriteSingleInfoFile(infoFile)
-	require.NoError(t, err)
-
-	// Verify content was written
-	content, err := afero.ReadFile(fs, "/test/.info")
-	require.NoError(t, err)
-	expected := "# Comment\nfile1.txt Test annotation"
-	assert.Equal(t, expected, string(content))
 }
