+++ treex/info/infofileset_test.go
@@ -8,159 +8,88 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-// TestInfoFileSet_Pure tests InfoFileSet operations as pure functions without filesystem
-func TestInfoFileSet_Pure(t *testing.T) {
-	t.Run("Gather consolidates annotations correctly", func(t *testing.T) {
-		// Create InfoFiles manually (no filesystem)
-		infoFile1 := NewInfoFile(".info", "a.txt  Root annotation\nb.txt  Another root annotation")
-		infoFile2 := NewInfoFile("sub/.info", "local.txt  Local annotation")
+// TestInfoFileSetLoader tests single file loading functionality
+func TestInfoFileSetLoader_LoadSingleInfoFile(t *testing.T) {
+	fs := afero.NewMemMapFs()
+	infoFS := NewAferoInfoFileSystem(fs)
+	loader := NewInfoFileSetLoader(infoFS)
 
-		// Mock pathExists (pure function)
-		pathExists := func(path string) bool {
-			validPaths := map[string]bool{
-				"a.txt":         true,
-				"b.txt":         true,
-				"sub/local.txt": true,
-			}
-			return validPaths[path]
-		}
+	// Create a test .info file
+	content := `file1.txt Annotation 1
+file2.txt Annotation 2`
 
-		// Create InfoFileSet and test Gather (pure operation)
-		infoFileSet := NewInfoFileSet([]*InfoFile{infoFile1, infoFile2}, pathExists)
-		gathered := infoFileSet.Gather()
+	err := afero.WriteFile(fs, "/test/.info", []byte(content), 0644)
+	require.NoError(t, err)
 
-		// Verify results
-		assert.Len(t, gathered, 3)
-		assert.Contains(t, gathered, "a.txt")
-		assert.Contains(t, gathered, "b.txt")
-		assert.Contains(t, gathered, "sub/local.txt")
+	// Load the file
+	infoFile, err := loader.LoadSingleInfoFile("/test/.info")
+	require.NoError(t, err)
 
-		// Root .info file annotation for a.txt (only annotation)
-		assert.Equal(t, "Root annotation", gathered["a.txt"].Annotation)
-		assert.Equal(t, ".info", gathered["a.txt"].InfoFile)
-	})
+	assert.Equal(t, "/test/.info", infoFile.Path)
+	assert.Len(t, infoFile.annotations, 2)
+	assert.True(t, infoFile.HasAnnotationForPath("file1.txt"))
+	assert.True(t, infoFile.HasAnnotationForPath("file2.txt"))
+}
 
-	t.Run("Validate detects cross-file conflicts", func(t *testing.T) {
-		infoFile1 := NewInfoFile(".info", "sub/target.txt  Root annotation")
-		infoFile2 := NewInfoFile("sub/.info", "target.txt  Sub annotation")
+// TestInfoFileSetLoader_LoadFromPath tests directory traversal and loading
+func TestInfoFileSetLoader_LoadFromPath(t *testing.T) {
+	fs := afero.NewMemMapFs()
+	infoFS := NewAferoInfoFileSystem(fs)
+	loader := NewInfoFileSetLoader(infoFS)
 
-		pathExists := func(path string) bool { return true }
+	// Create test .info files
+	err := afero.WriteFile(fs, "/project/.info", []byte("root.txt Root annotation"), 0644)
+	require.NoError(t, err)
 
-		infoFileSet := NewInfoFileSet([]*InfoFile{infoFile1, infoFile2}, pathExists)
-		result := infoFileSet.Validate()
+	err = fs.MkdirAll("/project/sub", 0755)
+	require.NoError(t, err)
 
-		// Should detect cross-file conflict
-		assert.True(t, len(result.Issues) > 0)
-		assert.True(t, infoFileSet.HasConflicts())
+	err = afero.WriteFile(fs, "/project/sub/.info", []byte("sub.txt Sub annotation"), 0644)
+	require.NoError(t, err)
 
-		// Find the cross-file conflict issue
-		var foundConflict bool
-		for _, issue := range result.Issues {
-			if issue.Type == IssueMultipleFiles {
-				foundConflict = true
-				assert.Equal(t, "sub/.info", issue.InfoFile) // Later file should be flagged
-				assert.Equal(t, ".info", issue.RelatedFile)  // Earlier file should be related
-			}
-		}
-		assert.True(t, foundConflict, "Should find cross-file conflict")
-	})
+	// Load all files
+	infoFileSet, err := loader.LoadFromPath("/project")
+	require.NoError(t, err)
 
-	t.Run("Clean removes problematic annotations", func(t *testing.T) {
-		// Create InfoFile with problematic annotations
-		infoFile := NewInfoFile(".info", `
-valid.txt     Valid annotation
-missing.txt   Annotation for missing file
-invalid_line
-duplicate.txt First annotation
-duplicate.txt Duplicate annotation
-`)
+	infoFiles := infoFileSet.GetFiles()
+	assert.Len(t, infoFiles, 2)
 
-		pathExists := func(path string) bool {
-			return path == "valid.txt" // Only valid.txt exists
-		}
+	// Check that both files were loaded
+	paths := make(map[string]bool)
+	for _, infoFile := range infoFiles {
+		paths[infoFile.Path] = true
+	}
+	assert.True(t, paths["/project/.info"])
+	assert.True(t, paths["/project/sub/.info"])
+}
 
-		infoFileSet := NewInfoFileSet([]*InfoFile{infoFile}, pathExists)
-		cleanResult, cleanedSet := infoFileSet.Clean()
+// TestInfoFileSetWriter_WriteSingleInfoFile tests single file writing
+func TestInfoFileSetWriter_WriteSingleInfoFile(t *testing.T) {
+	fs := afero.NewMemMapFs()
+	infoFS := NewAferoInfoFileSystem(fs)
+	writer := NewInfoFileSetWriter(infoFS)
 
-		// Verify clean results
-		assert.True(t, cleanResult.Summary.FilesModified > 0)
-		assert.True(t, cleanResult.Summary.InvalidPathsRemoved > 0)
-		assert.Len(t, cleanResult.RemovedAnnotations, 2) // missing.txt + duplicate
+	// Create directory first
+	err := fs.MkdirAll("/test", 0755)
+	require.NoError(t, err)
 
-		// Verify cleaned set only has valid annotation
-		cleanedAnnotations := cleanedSet.GetAllAnnotations()
-		assert.Len(t, cleanedAnnotations, 1)
-		assert.Equal(t, "valid.txt", cleanedAnnotations[0].Path)
-	})
+	// Create an InfoFile with some content to avoid empty string issue
+	infoFile := NewInfoFile("/test/.info", "# Comment")
+	infoFile.AddAnnotationForPath("file1.txt", "Test annotation")
 
-	t.Run("Filter and RemoveEmpty work correctly", func(t *testing.T) {
-		emptyFile := NewInfoFile("empty/.info", "")
-		nonEmptyFile := NewInfoFile("full/.info", "file.txt  Annotation")
+	// Write to disk
+	err = writer.WriteSingleInfoFile(infoFile)
+	require.NoError(t, err)
 
-		pathExists := func(path string) bool { return true }
-
-		infoFileSet := NewInfoFileSet([]*InfoFile{emptyFile, nonEmptyFile}, pathExists)
-
-		// Test Filter
-		filtered := infoFileSet.Filter(func(file *InfoFile) bool {
-			return file.Path == "full/.info"
-		})
-		assert.Equal(t, 1, filtered.Count())
-		assert.Equal(t, "full/.info", filtered.GetFiles()[0].Path)
-
-		// Test RemoveEmpty
-		nonEmpty := infoFileSet.RemoveEmpty()
-		assert.Equal(t, 1, nonEmpty.Count())
-		assert.Equal(t, "full/.info", nonEmpty.GetFiles()[0].Path)
-
-		// Test GetEmptyFiles
-		emptyFiles := infoFileSet.GetEmptyFiles()
-		assert.Len(t, emptyFiles, 1)
-		assert.Equal(t, "empty/.info", emptyFiles[0].Path)
-	})
-
-	t.Run("Distribute moves annotations to optimal files", func(t *testing.T) {
-		// Create scenario where annotation in root should move to sub directory
-		rootFile := NewInfoFile(".info", "sub/target.txt  Root annotation")
-		subFile := NewInfoFile("sub/.info", "local.txt  Local annotation")
-
-		pathExists := func(path string) bool { return true }
-
-		infoFileSet := NewInfoFileSet([]*InfoFile{rootFile, subFile}, pathExists)
-		distributed := infoFileSet.Distribute()
-
-		// Verify annotation moved to closer file and empty files are removed
-		files := distributed.GetFiles()
-		var distributedSubFile *InfoFile
-		for _, file := range files {
-			if file.Path == "sub/.info" {
-				distributedSubFile = file
-			}
-		}
-
-		// Root file should be removed entirely after distribution (was empty)
-		assert.Len(t, files, 1, "Only sub/.info should remain after distribution")
-		require.NotNil(t, distributedSubFile, "Sub file should exist")
-
-		subAnnotations := distributedSubFile.GetAllAnnotations()
-		assert.Len(t, subAnnotations, 2, "Sub file should have local.txt + target.txt")
-
-		// Verify the specific annotations
-		var hasLocal, hasTarget bool
-		for _, ann := range subAnnotations {
-			if ann.Path == "local.txt" {
-				hasLocal = true
-			}
-			if ann.Path == "target.txt" { // Should be relative to sub/.info
-				hasTarget = true
-			}
-		}
-		assert.True(t, hasLocal, "Should have local.txt annotation")
-		assert.True(t, hasTarget, "Should have target.txt annotation (moved from root)")
-	})
+	// Verify content was written
+	content, err := afero.ReadFile(fs, "/test/.info")
+	require.NoError(t, err)
+	expected := "# Comment\nfile1.txt Test annotation"
+	assert.Equal(t, expected, string(content))
 }
 
 // TestInfoFileSet_WithFilesystem tests filesystem I/O boundaries
+// These tests are extracted from the working filesystem integration tests
 func TestInfoFileSet_WithFilesystem(t *testing.T) {
 	t.Run("LoadFromPath reads all .info files", func(t *testing.T) {
 		// Create in-memory filesystem
@@ -263,126 +192,66 @@ b.txt  Second annotation
 	})
 }
 
-// TestInfoFileSet_MethodChaining tests fluent interface
-func TestInfoFileSet_MethodChaining(t *testing.T) {
-	t.Run("Chain operations fluently", func(t *testing.T) {
-		// Create problematic InfoFiles
-		emptyFile := NewInfoFile("empty/.info", "")
-		invalidFile := NewInfoFile("invalid/.info", "missing.txt  Annotation for missing file")
-		validFile := NewInfoFile("valid/.info", "exists.txt  Valid annotation")
+// TestAferoInfoFileSystem tests the filesystem abstraction
+func TestAferoInfoFileSystem(t *testing.T) {
+	fs := afero.NewMemMapFs()
+	infoFS := NewAferoInfoFileSystem(fs)
 
-		pathExists := func(path string) bool {
-			// Only valid/exists.txt exists (resolved paths)
-			// invalid/missing.txt does not exist
-			return path == "valid/exists.txt"
-		}
+	t.Run("ReadInfoFile", func(t *testing.T) {
+		content := "test.txt Test annotation"
+		err := afero.WriteFile(fs, ".info", []byte(content), 0644)
+		require.NoError(t, err)
 
-		infoFileSet := NewInfoFileSet([]*InfoFile{emptyFile, invalidFile, validFile}, pathExists)
+		reader, err := infoFS.ReadInfoFile(".info")
+		require.NoError(t, err)
 
-		// Chain operations: Clean -> RemoveEmpty -> Filter
-		_, cleanedSet := infoFileSet.Clean()
-		result := cleanedSet.
-			RemoveEmpty().
-			Filter(func(file *InfoFile) bool {
-				return len(file.GetAllAnnotations()) > 0
-			})
-
-		// Should only have the valid file with valid annotations
-		assert.Equal(t, 1, result.Count())
-		annotations := result.GetAllAnnotations()
-		assert.Len(t, annotations, 1)
-		assert.Equal(t, "exists.txt", annotations[0].Path)
+		data, err := afero.ReadAll(reader)
+		require.NoError(t, err)
+		assert.Equal(t, content, string(data))
 	})
 
-	t.Run("WithPathValidator changes validation behavior", func(t *testing.T) {
-		infoFile := NewInfoFile(".info", "missing.txt  Annotation")
+	t.Run("WriteInfoFile", func(t *testing.T) {
+		content := "file.txt File annotation"
+		err := infoFS.WriteInfoFile("test/.info", content)
+		require.NoError(t, err)
 
-		// Initially, missing.txt doesn't exist
-		pathExists1 := func(path string) bool { return false }
-		infoFileSet := NewInfoFileSet([]*InfoFile{infoFile}, pathExists1)
-
-		result1 := infoFileSet.Validate()
-		assert.True(t, len(result1.Issues) > 0) // Should have issues
-
-		// Change validator to make missing.txt exist
-		pathExists2 := func(path string) bool { return true }
-		newSet := infoFileSet.WithPathValidator(pathExists2)
-
-		result2 := newSet.Validate()
-		assert.True(t, len(result2.Issues) == 0) // Should have no issues now
-	})
-}
-
-// TestInfoFileSet_EdgeCases tests edge cases and error conditions
-func TestInfoFileSet_EdgeCases(t *testing.T) {
-	t.Run("Empty InfoFileSet operations", func(t *testing.T) {
-		pathExists := func(path string) bool { return true }
-		emptySet := EmptyInfoFileSet(pathExists)
-
-		assert.Equal(t, 0, emptySet.Count())
-		assert.Equal(t, 0, emptySet.GetAnnotationCount())
-		assert.False(t, emptySet.HasConflicts())
-
-		gathered := emptySet.Gather()
-		assert.Len(t, gathered, 0)
-
-		distributed := emptySet.Distribute()
-		assert.Equal(t, 0, distributed.Count())
-
-		result := emptySet.Validate()
-		assert.Len(t, result.Issues, 0)
-		assert.Equal(t, 0, result.Summary["total_files"])
+		data, err := afero.ReadFile(fs, "test/.info")
+		require.NoError(t, err)
+		assert.Equal(t, content, string(data))
 	})
 
-	t.Run("InfoFileSet with nil pathExists", func(t *testing.T) {
-		infoFile := NewInfoFile(".info", "test.txt  Test annotation")
-		infoFileSet := NewInfoFileSet([]*InfoFile{infoFile}, nil)
+	t.Run("PathExists", func(t *testing.T) {
+		err := afero.WriteFile(fs, "existing.txt", []byte("content"), 0644)
+		require.NoError(t, err)
 
-		// Should work without path validation
-		gathered := infoFileSet.Gather()
-		assert.Len(t, gathered, 1)
-
-		// Validation should skip path existence checks
-		result := infoFileSet.Validate()
-		hasPathNotExistsIssues := false
-		for _, issue := range result.Issues {
-			if issue.Type == IssuePathNotExists {
-				hasPathNotExistsIssues = true
-				break
-			}
-		}
-		assert.False(t, hasPathNotExistsIssues, "Should not check path existence with nil pathExists")
+		assert.True(t, infoFS.PathExists("existing.txt"))
+		assert.False(t, infoFS.PathExists("nonexistent.txt"))
 	})
 
-	t.Run("Complex precedence rules", func(t *testing.T) {
-		// Test complex precedence scenario with multiple levels
-		files := []*InfoFile{
-			NewInfoFile(".info", "rootfile.txt  Root level"),
-			NewInfoFile("target/.info", "file.txt  Target level"),
-			NewInfoFile("target/sub/.info", "subfile.txt  Sub level"),
-		}
+	t.Run("FindInfoFiles", func(t *testing.T) {
+		// Use a clean filesystem for this test
+		testFS := afero.NewMemMapFs()
+		testInfoFS := NewAferoInfoFileSystem(testFS)
 
-		pathExists := func(path string) bool { return true }
-		infoFileSet := NewInfoFileSet(files, pathExists)
+		// Create .info files in different directories
+		err := afero.WriteFile(testFS, "test/.info", []byte("root"), 0644)
+		require.NoError(t, err)
 
-		gathered := infoFileSet.Gather()
+		err = testFS.MkdirAll("test/sub/deep", 0755)
+		require.NoError(t, err)
 
-		// Each annotation should be present since they don't conflict
-		assert.Len(t, gathered, 3)
+		err = afero.WriteFile(testFS, "test/sub/.info", []byte("sub"), 0644)
+		require.NoError(t, err)
 
-		// Root level annotation
-		rootWinner := gathered["rootfile.txt"]
-		assert.Equal(t, "Root level", rootWinner.Annotation)
-		assert.Equal(t, ".info", rootWinner.InfoFile)
+		err = afero.WriteFile(testFS, "test/sub/deep/.info", []byte("deep"), 0644)
+		require.NoError(t, err)
 
-		// Target level annotation
-		targetWinner := gathered["target/file.txt"]
-		assert.Equal(t, "Target level", targetWinner.Annotation)
-		assert.Equal(t, "target/.info", targetWinner.InfoFile)
+		infoFiles, err := testInfoFS.FindInfoFiles("test")
+		require.NoError(t, err)
 
-		// Sub level annotation
-		subWinner := gathered["target/sub/subfile.txt"]
-		assert.Equal(t, "Sub level", subWinner.Annotation)
-		assert.Equal(t, "target/sub/.info", subWinner.InfoFile)
+		assert.Len(t, infoFiles, 3)
+		assert.Contains(t, infoFiles, "test/.info")
+		assert.Contains(t, infoFiles, "test/sub/.info")
+		assert.Contains(t, infoFiles, "test/sub/deep/.info")
 	})
 }
